refactor rails specific logic into separate files for clarify, add additional specs
