+++ spec/integration/view/view_additions_rails6_spec.rb
@@ -1,6 +1,6 @@
 require './spec/rails_helper'
 
-if ActionPack::VERSION::MAJOR < 7
+if ActionPack::VERSION::MAJOR == 6
   describe ActionView::PathResolver do
     let(:resolver) { ActionView::PathResolver.new }
